+++ src/pages/SettingsPage.component.jsx
@@ -9,7 +9,7 @@ function SettingsPage() {
     document.title = "Settings Page - Kitchen Inventory";
   }, [])
   return (
-    <main>
+    <article>
       <h2>Settings</h2>
       <p>
         This is the settings page. Here you can customize your preferences and
@@ -25,7 +25,7 @@ function SettingsPage() {
       </button>
 
       {showFieldSelector && <FieldSelector onClose={closeFieldSelector} />}
-    </main>
+    </article>
   );
 }